+++ components/sections/Analytics.tsx
@@ -1,7 +1,0 @@
-import { analyticsRepository } from "@/repositories"
-import AnalyticsClient from "./AnalyticsClient"
-
-export default async function Analytics() {
-  const data = await analyticsRepository.get()
-  return <AnalyticsClient data={data} />
-}